+++ source/server/modules/ida/readIdaItem.c
@@ -2733,19 +2733,19 @@ int readIdaItem(char* itemname, ida_file_ptr* ida_file, short* context, DATA_BLO
     if (data_block->client_block.get_dimdble || data_block->client_block.get_timedble) {
         float* foffs, * fints;
         double* doffs = NULL, * dints = NULL;
-        for (i = 0; i < data_block->rank; i++) {
+        for (i = 0; i < (int)data_block->rank; i++) {
             if (data_block->client_block.get_dimdble ||
                 (data_block->client_block.get_timedble && i == data_block->order)) {
                 data_block->dims[i].data_type = UDA_TYPE_DOUBLE;
                 if ((foffs = (float*)data_block->dims[i].offs) != NULL) {
                     doffs = (double*)realloc((void*)doffs, data_block->dims[i].udoms * sizeof(double));
-                    for (j = 0; j < data_block->dims[i].udoms; j++) doffs[j] = (double)foffs[j];
+                    for (j = 0; j < (int)data_block->dims[i].udoms; j++) doffs[j] = (double)foffs[j];
                     free((void*)data_block->dims[i].offs);
                     data_block->dims[i].offs = (char*)doffs;
                 }
                 if ((fints = (float*)data_block->dims[i].ints) != NULL) {
                     dints = (double*)realloc((void*)dints, data_block->dims[i].udoms * sizeof(double));
-                    for (j = 0; j < data_block->dims[i].udoms; j++) dints[j] = (double)fints[j];
+                    for (j = 0; j < (int)data_block->dims[i].udoms; j++) dints[j] = (double)fints[j];
                     free((void*)data_block->dims[i].ints);
                     data_block->dims[i].ints = (char*)dints;
                 }
